Add ability to play in browser
This is extremely basic right now and the interface is prettttty bad.
Create a game with the "new game" button
copy the page url and replace the word "play" in the path with "join"
send that link to somebody else.
now that you've both joined, you will see the FEN version of the board state, and can submit
moves in a text box.
There's no indication of whose turn it is or anything.