Add ability to resign from a game
This was a bit awkward because the library i'm using for maintaining the chess game
state, binbo, doesn't handle resignations. I've been using the binbo process as the
single source of truth for game state, and with this change that isn't true anymore.

I just bolted a "winner" field onto the response and if that is set, and the game status
is continue, then i assume resignation.
I think the nicer way to do this is to add it to the status field, instead of having that
field come straight from binbo.