Don't show game until both players have connected
I added a bool flag to the game state returned by the backend:
"haveBothPlayersConnected"

I'm still wondering about the best way to model this. I'd rather have
a tuple kind of thing that doesnt even send game state if we both havent
connected yet, but you can't represent that in JSON directly, so i'd have
to change the whole format.
The main reason i'd represent it differently is the flag is meaningless after
the initial connection.
Eventually I'd like to handle player disconnections though so might be able to
use this for that.