Allow send promotions other than queen
I needed some server/api changes because, once i started sending promotions, I found out that
they weren't handled quite right on the backend. The main change is expecting moves to come
with promotion = null instead of no promotion field.

The UI is pretty basic. When you get a pawn to the end of the board, a list of available pieces to
promote to displays in the middle of the game board. Click on one of these pieces to select it,
or click outside of the promotion menu to cancel and pick another move.